R6 Peak Region Finder

R6 Peak Region Finder

R6 Peak Region Finder

Details

Think of it like managing all the stuff that needs to happen to find the peaks in the regions.

Public fields

run_time

how long did the process take

start_time

when did we start

stop_time

when did we start

peak_regions

SCPeakRegions object

sliding_region_size

how big are the sliding regions in data points

sliding_region_delta

how much space between sliding region starts

quantile_multiplier

how much to multiply quantile based cutoff by

n_point_region

how many points are there in the big tiled regions for quantile based cutoff

tiled_region_size

how wide are the tiled regions in data points

tiled_region_delta

how far in between each tiled region

region_percentile

??

peak_method

what method to extract peak center, height, area, etc

min_points

how many points wide does a peak have to be to get characterized

sample_id

what sample are we processing

zero_normalization

do we want to pretend to do normalization

Methods

Public methods

SCPeakRegionFinder$add_regions()

SCPeakRegionFinder$reduce_sliding_regions()

SCPeakRegionFinder$split_peak_regions()

SCPeakRegionFinder$remove_double_peaks_in_scans()

SCPeakRegionFinder$normalize_data()

SCPeakRegionFinder$find_peaks_in_regions()

SCPeakRegionFinder$model_mzsd()

SCPeakRegionFinder$model_heightsd()

SCPeakRegionFinder$indicate_high_frequency_sd()

SCPeakRegionFinder$add_data()

SCPeakRegionFinder$summarize_peaks()

SCPeakRegionFinder$add_offset()

SCPeakRegionFinder$sort_ascending_mz()

SCPeakRegionFinder$characterize_peaks()

SCPeakRegionFinder$summarize()

SCPeakRegionFinder$peak_meta()

SCPeakRegionFinder$new()

SCPeakRegionFinder$clone()

Method add_regions()

Add the sliding and tiled regions

Usage

SCPeakRegionFinder$add_regions()

Method reduce_sliding_regions()

Find the regions most likely to contain real signal

Usage

SCPeakRegionFinder$reduce_sliding_regions()

Method split_peak_regions()

Split up signal regions by peaks found

Usage

SCPeakRegionFinder$split_peak_regions(use_regions = NULL)

Arguments

use_regions

an index of the regions we want to split up

Method remove_double_peaks_in_scans()

Check for the presence of two peaks with the same scan number in each region and remove them. Any regions with zero peaks left, remove the region.

Usage

SCPeakRegionFinder$remove_double_peaks_in_scans()

Method normalize_data()

Normalize the intensity data

Usage

SCPeakRegionFinder$normalize_data(which_data = "both")

Arguments

which_data

raw, characterized, or both (default)

Method find_peaks_in_regions()

Find the peaks in the regions.

Usage

SCPeakRegionFinder$find_peaks_in_regions()

Method model_mzsd()

Model the m/z standard deviation.

Usage

SCPeakRegionFinder$model_mzsd()

Method model_heightsd()

Model the intensity height standard deviation.

Usage

SCPeakRegionFinder$model_heightsd()

Method indicate_high_frequency_sd()

Look for peaks with higher than expected frequency standard deviation.

Usage

SCPeakRegionFinder$indicate_high_frequency_sd()

Method add_data()

Add the data from an SCMzml object to the underlying SCPeakRegions object.

Usage

SCPeakRegionFinder$add_data(sc_mzml)

Arguments

sc_mzml

the SCMzml object being passed in

Method summarize_peaks()

Summarize the peaks to go into JSON form.

Usage

SCPeakRegionFinder$summarize_peaks()

Method add_offset()

Add an offset based on width in frequency space to m/z to describe how wide the peak is.

Usage

SCPeakRegionFinder$add_offset()

Method sort_ascending_mz()

Sort the data in m/z order, as the default is frequency order

Usage

SCPeakRegionFinder$sort_ascending_mz()

Method characterize_peaks()

Run the overall peak characterization from start to finish.

Usage

SCPeakRegionFinder$characterize_peaks()

Method summarize()

Summarize everything for output to the zip file after completion.

Usage

SCPeakRegionFinder$summarize(
  package_used = "package:ScanCentricPeakCharacterization"
)

Arguments

package_used

which package is being used for this work.

Method peak_meta()

Generate the meta data that goes into the accompanying JSON file.

Usage

SCPeakRegionFinder$peak_meta()

Method new()

Make a new SCPeakRegionFinder object.

Usage

SCPeakRegionFinder$new(
  sc_mzml = NULL,
  sliding_region_size = 10,
  sliding_region_delta = 1,
  tiled_region_size = 1,
  tiled_region_delta = 1,
  region_percentile = 0.99,
  offset_multiplier = 1,
  frequency_multiplier = 400,
  quantile_multiplier = 1.5,
  n_point_region = 2000,
  peak_method = "lm_weighted",
  min_points = 4,
  zero_normalization = FALSE
)

Arguments

sc_mzml

the SCMzml object to use (can be missing)

sliding_region_size

how wide to make the sliding regions in data points

sliding_region_delta

how far apart are the starting locations of the sliding regions

tiled_region_size

how wide are the tiled regions

tiled_region_delta

how far apart are the tiled reigons

region_percentile

cumulative percentile cutoff to use

offset_multiplier

what offset multiplier should be used

frequency_multiplier

how much to multiply frequency points to interval ranges

quantile_multiplier

how much to adjust the quantile cutoff by

n_point_region

how many points in the large tiled regions

peak_method

the peak characterization method to use (lm_weighted)

min_points

how many points to say there is a peak (4)

zero_normalization

don't actually do normalization (FALSE)

Method clone()

The objects of this class are cloneable with this method.

Usage

SCPeakRegionFinder$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.