Holds all the peak region data

Holds all the peak region data

Holds all the peak region data

Details

This reference class represents the peak region data.

Public fields

frequency_point_regions

the frequency data

frequency_fit_description

the model of frequency ~ m/z

mz_fit_description

the model of m/z ~ frequency

peak_regions

the peak regions

sliding_regions

the sliding regions used for density calculations

tiled_regions

the tiled regions used for grouping and splitting peak regions

peak_region_list

list of regions

frequency_multiplier

how much to multiplier frequency by to make interval points

scan_peaks

the peaks by scans

peak_data

the data.frame of final peak data

scan_level_arrays

scan level peak data as matrices

is_normalized

are the peak intensities normalized

normalization_factors

the normalization factors calculated

n_scan

how many scans are we working with

scans_per_peak

??

scan_perc

what percentage of scans is a minimum

min_scan

based on scan_perc, how many scans minimum does a peak need to be in

max_subsets

??

scan_subsets

??

frequency_range

what is the range in frequency space

scan_correlation

??

keep_peaks

which peaks are we keeping out of all the peaks we had

peak_index

the indices for the peaks

scan_indices

the names of the scans

instrument

the instrument serial number if available

Methods

Public methods

SCPeakRegions$set_min_scan()

SCPeakRegions$add_data()

SCPeakRegions$new()

SCPeakRegions$clone()

Method set_min_scan()

sets the minimum number of scans to use

Usage

SCPeakRegions$set_min_scan()

Method add_data()

Adds the data from an SCMzml object to the SCPeakRegion.

Usage

SCPeakRegions$add_data(sc_mzml)

Arguments

sc_mzml

the SCMzml object being passed in

Method new()

Creates a new SCPeakRegions object

Usage

SCPeakRegions$new(
  sc_mzml = NULL,
  frequency_multiplier = 400,
  scan_perc = 0.1,
  max_subsets = 100
)

Arguments

sc_mzml

the SCMzml object to get data from

frequency_multiplier

how much to multiply frequency by

scan_perc

how many scans are required to be in to be a "peak"

max_subsets

??

Method clone()

The objects of this class are cloneable with this method.

Usage

SCPeakRegions$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.